Fixed camera u and v ray cast parameters

+++ include/mcrt/Camera.h
@@ -20,8 +20,8 @@ namespace Mcrt
                 float y,
                 uint16_t width,
                 uint16_t height,
-                uint16_t u,
-                uint16_t v
+                float u,
+                float v
         ) const;
 
         glm::vec3 get_pos() const;
